+++ scopes/explorer/ui/component-card/component-card.tsx
@@ -52,7 +52,6 @@ export type ComponentCardProps = {
    */
   href?: string;
   external?: boolean;
-  hidePreview?: boolean;
 } & BaseComponentCardProps;
 
 export function ComponentCard({
@@ -65,13 +64,12 @@ export function ComponentCard({
   isDeprecated = false,
   href,
   external,
-  hidePreview,
 }: ComponentCardProps) {
   return (
     <Card className={className}>
       <Link className={styles.componentCardLink} href={href || id} external={external}>
         <DeprecationSticker isDeprecated={isDeprecated} />
-        {hidePreview || <PreviewContainer preview={preview} />}
+        <PreviewContainer preview={preview} />
         <ComponentDetails id={id} version={version} description={description} className={styles.content} />
         <div className={styles.bottom}>
           <div className={styles.left} />
