+++ scopes/lanes/ui/hooks/index.ts
@@ -11,4 +11,6 @@ export {
   LanesHost,
   LanesQuery,
 } from './lanes-context';
-export { useLaneReadme, useLaneComponents, useLanes } from './lanes-query';
+export { useLaneComponents } from './use-lane-components';
+export { useLaneReadme } from './use-lane-readme';
+export { useLanes } from './use-lanes';
