--- scopes/lanes/ui/hooks/lanes-query/index.ts
@@ -1,3 +1,0 @@
-export { useLaneComponents } from './use-lane-components';
-export { useLaneReadme } from './use-lane-readme';
-export { useLanes } from './use-lanes';
